requirements.txt
git+https://github.com/atkinssamuel/google-auth-wrapper.git
feedparser
openai
python-dotenv
google-api-python-client
